monitor/requirements.txt
paho-mqtt
psycopg2
sqlalchemy